Create an inventory and config
Even though it's unlikely that I'll need to run this against more hosts than
just localhost it still makes sense to have the host list defined in a custom
invetory. sakura.yml is growing large enough already that I'd like to split it
out into multiple playbooks, and specifying host: localhost at the top of every
one is a recipe for confusion later on.