Only use become where necessary
Previously we were installing all gnome extensions as root. This is because we
were using become to escalate permissions for the entire playbook, which meant
every task was running as root. We should only escalate privs where necessary,
which makes the playbook a little more verbose, but at least we're being
explicit about our intentions