Fix gnome extensions.
Gnome extensions defines a var in the role - which it turns out is really high
up in the variable priority chain. The only realistic way to override it is to
define a role parameter (not a role variable) or pass args on the command line.

This doesn't seem optimal as it means we can't define our own extension list
using vars, and we also can't define our own list per host.